==> step-0-import-and-clean.ipynb <==
### Import and Clean Tweets

import pandas as pd
import numpy as np

tweets = pd.read_csv( "data/tweets.csv", usecols=[ "text", "created_at" ] )
tweets.head()

# create year & month bins, and downcast to int32
tweets.created_at = pd.to_datetime( tweets.created_at, infer_datetime_format=True, errors='coerce' )

tweets[ "year" ] = tweets.created_at.dt.year
tweets[ "month" ] = tweets.created_at.dt.month
tweets[ "day" ] = tweets.created_at.dt.day
tweets[ "hour" ] = tweets.created_at.dt.hour
tweets[ "minute" ] = tweets.created_at.dt.minute
#tweets[ "time"] = tweets.created_at.dt.time
#tweets[ "date" ] = tweets.created_at.dt.date
#tweets[ "hour_f" ] = ( tweets.created_at.dt.hour * 1.0 ) + tweets.created_at.dt.minute / 60.0

# reduce storage size
tweets.month = tweets.month.astype( np.int32 )
tweets.year  = tweets.year.astype( np.int32 )
tweets.day   = tweets.day.astype( np.int32 )
tweets.hour   = tweets.hour.astype( np.int32 )
tweets.minute = tweets.minute.astype( np.int32 )
#tweets.hour_f  = tweets.hour.astype( np.float32 )

#tweets.drop( ['created_at'] , axis=1, inplace=True )

tweets.info()

tweets.count()

tweets.dropna( inplace=True )

tweets.count()

tweets.count()

# how many per year?
tweets[ 'year' ].value_counts()

# how many by hour?
tweets[ 'hour' ].value_counts()

#tweets.groupby( [ "year", "month" ] ).count().reset_index()
#tweets.groupby( [ "year" ] ).count().reset_index()

tweets.groupby( ['year', 'month'] )[ 'month' ].count()

tweets.head()

tweets.to_csv( "data/tweets-clean.csv", index=False )
