==> step-2-gensim-toaster.ipynb <==
## Gensim Toaster

# Used to import other notebooks: https://pypi.org/project/import-ipynb/
import import_ipynb
import deepily.util as dp_util
from gensim.summarization import summarize, summarize_corpus, keywords
from gensim.summarization.textcleaner import clean_text_by_sentences

# http://cmdlinetips.com/2011/08/three-ways-to-read-a-text-file-line-by-line-in-python/
def load_doc_by_line( filename ):
    
    # Open the file with read only permit
    file = open( filename, "r" )
    
    # use readlines to read all lines in the file
    # The variable "lines" is a list containing all lines in the file
    lines = file.readlines()
    
    # close the file after reading the lines.
    file.close()
    
    return lines

start_time = dp_util.get_time()
in_filename = "../texts/trump-tweets.txt"
tweets = load_doc_by_line( in_filename )
dp_util.print_time( start_time, dp_util.get_time() )

print( len( tweets ) )
tweets[ 0 ]

### Extract Key Words from Tweet

tweet = tweets[ 1 ]
print( tweet )

keywords_list = keywords( tweet ).split( "\n" )

for word in keywords_list: 
    print( word )

### Get Summary --or-- Keywords If Unable to Summarize

for tweet in tweets[ :10 ]:
    
    print( "\n---------------------------------" )
        
    if len( clean_text_by_sentences( tweet ) ) < 2:
        print( "TOO SHORT TO SUMMARIZE:", tweet )
    else:
        print( "Original:", tweet )
        print()
        summary = summarize( tweet, ratio=0.5 )
        if summary != '':
            print( "Summary:", summary )
        else:
            print( "keywords:", keywords( tweet ).split( "\n" ) )
