==> step-2-spacy-toaster.ipynb <==
import spacy
import pickle
from spacy import displacy

tweets_by_year = None
tweets_by_month = None

tweets_month_file_name = 'data/tweets-by-month.data'
tweets_window_file_name = 'data/tweets-by-window.data'

with open( tweets_window_file_name, 'rb' ) as filehandle:  
    
    # read the data as binary data stream
    tweets_by_year = pickle.load( filehandle )
    
with open( tweets_month_file_name, 'rb' ) as filehandle:  
    
    # read the data as binary data stream
    tweets_by_month = pickle.load( filehandle )
    
print( len( tweets_by_year ) )
print( len( tweets_by_month ) )

nlp = spacy.load( 'en' )

### Linguistic annotations

doc = u'Apple is looking at buying U.K. startup for $1 billion'

def print_pos( doc ):
    
    doc = nlp( doc )
    
    for token in doc:
        print( token.text, token.pos_, token.dep_ )
        
print_pos( doc )        

last_month = tweets_by_month[ -1 ]

print( last_month[ 0 ][ -1 ], last_month[ 1 ][ -1 ] )

tweet = last_month[ 1 ][ -1 ]
print_pos( tweet )

### Tokenization

def print_tokens( doc ):
    
    doc = nlp( doc )
    for token in doc:
        print( token.text )
        
tweet = last_month[ 1 ][ -1 ]
print_tokens( tweet )        

### PoS + Dependencies

def print_pos_and_dependencies( doc ):

    doc = nlp( doc )
    for token in doc:
        print( token.text, token.lemma_, token.pos_, token.tag_, token.dep_, token.shape_, token.is_alpha, token.is_stop )
        
print_pos_and_dependencies( tweet )        

### Named Entities

def print_named_entities( doc ):
    
    doc = nlp( doc )
    
    for ent in doc.ents:
        print( ent.text, ent.start_char, ent.end_char, ent.label_ )
    
print_named_entities( tweet )    

### Visualize Parse

# from: https://stackoverflow.com/questions/25698448/how-to-embed-html-into-ipython-output
from IPython.core.display import display, HTML

#doc = nlp( u'This is a sentence.' )
display( HTML( displacy.render( nlp( tweet ), style='dep', page="true" ) ) )

### Word Vectors and Similarity

tokens = u'dog cat banana'

def print_similiary( tokens ):
    
    tokens = nlp( tokens )
    for token1 in tokens:
        for token2 in tokens:
            print( token1.text, token2.text, token1.similarity( token2 ) )
            
print_similiary( tokens )            

print_similiary( "rocks mountains" )

print_similiary( "politician human" )
